// src/materials/enums.ts
/**
 * 枚举类型
 */

export enum MaterialGroup {
  Base = 'base',
  Media = 'media',
  Chart = 'chart',
  Market = 'market',
}

export enum FormItemType {
  Input = 'Input',
  InputNumber = 'InputNumber',
  Select = 'Select',
  Switch = 'Switch',
  ColorPicker = 'ColorPicker',
  DatePicker = 'DatePicker',
  PageSelector = 'PageSelector',
  MultiplePageSelector = 'MultiplePageSelector',
  IconSelector = 'IconSelector',
  VideoSelector = 'VideoSelector',
  AudioSelector = 'AudioSelector',
}